Referencias bigInteger a colunas integer
Substituicao das FK bigInteger por FK integer
pois seus alvos sao colunas integer

+++ migrations/20150422195549_cidade.js
@@ -5,7 +5,7 @@ exports.up = function(knex, Promise) {
     table.increments('id_cidade').primary();
     table.string('nome').notNullable();
     table.boolean('ativo').notNullable();
-    table.bigInteger('estado_id').notNullable()
+    table.integer('estado_id').notNullable()
       .references('id_estado').inTable('estado');
   });
 };
