+++ migrations/20150422195557_revendedor.js
@@ -3,7 +3,7 @@
 exports.up = function(knex, Promise) {
   return knex.schema
         .createTable('revendedor', function(table) {
-          table.bigInteger('pessoa_id')
+          table.integer('pessoa_id')
             .primary()
             .references('id_pessoa').inTable('pessoa');
           table.boolean('ativo').notNullable().defaultTo(true);
